+++ include/provenancePovJSON.h
@@ -13,8 +13,6 @@
 #ifndef __PROVENANCEPROVJSON_H
 #define __PROVENANCEPROVJSON_H
 
-int add_taint(const uint64_t id, const char* name);
-
 void set_ProvJSON_callback( void (*fcn)(char* json) );
 void flush_json( void );
 void append_activity(char* json_element);
@@ -52,7 +50,7 @@ char* packet_to_json(struct pck_struct* n);
 char* str_msg_to_json(struct str_struct* n);
 char* addr_to_json(struct address_struct* n);
 char* pathname_to_json(struct file_name_struct* n);
-char* prefix_json();
+const char* prefix_json();
 char* machine_description_json(char* buffer);
 char* iattr_to_json(struct iattr_prov_struct* n);
 char* xattr_to_json(struct xattr_prov_struct* n);