+++ include/provenanceutils.h
@@ -69,6 +69,8 @@ static const char RL_STR_RCV_PACKET []            = "receive_packet";
 static const char RL_STR_PERM_READ[]              = "perm_read";
 static const char RL_STR_PERM_WRITE[]             = "perm_write";
 static const char RL_STR_PERM_EXEC[]              = "perm_exec";
+static const char RL_STR_TERMINATE_PROCESS[]      = "terminate";
+static const char RL_STR_CLOSED[]						      = "closed";
 
 static inline const char* relation_str(uint64_t type){
   switch(type){
@@ -104,6 +106,8 @@ static inline const char* relation_str(uint64_t type){
       return RL_STR_SETXATTR;
     case RL_RMVXATTR:
       return RL_STR_RMVXATTR;
+    case RL_NAMED:
+      return RL_STR_NAMED;
     case RL_NAMED_PROCESS:
       return RL_STR_NAMED_PROCESS;
     case RL_EXEC:
@@ -146,6 +150,10 @@ static inline const char* relation_str(uint64_t type){
       return RL_STR_SH_WRITE;
     case RL_PERM_EXEC:
       return RL_STR_PERM_EXEC;
+		case RL_TERMINATE_PROCESS:
+			return RL_STR_TERMINATE_PROCESS;
+		case RL_CLOSED:
+			return RL_STR_CLOSED;
     default:
       return RL_STR_UNKNOWN;
   }
@@ -192,6 +200,8 @@ static inline const uint64_t relation_id(char* str){
   MATCH_AND_RETURN(str, RL_STR_PERM_EXEC, RL_PERM_EXEC);
   MATCH_AND_RETURN(str, RL_STR_SH_READ, RL_SH_READ);
   MATCH_AND_RETURN(str, RL_STR_SH_WRITE, RL_SH_WRITE);
+  MATCH_AND_RETURN(str, RL_STR_TERMINATE_PROCESS, RL_TERMINATE_PROCESS);
+  MATCH_AND_RETURN(str, RL_STR_CLOSED, RL_CLOSED);
   return 0;
 }