+++ service/audit.c
@@ -97,8 +97,7 @@ void log_disc(struct disc_node_struct* node){
     case AGT_DISC:
       append_agent(disc_to_json(node));
       break;
-    case ENT_DISC:
-    default:
+    default: // default to entity
       append_entity(disc_to_json(node));
       break;
   }
@@ -136,11 +135,7 @@ void log_packet_content(struct pckcnt_struct* cnt){
   append_entity(pckcnt_to_json(cnt));
 }
 
-bool filter(union prov_msg* msg){
-  return false;
-}
-
-bool long_filter(union long_prov_msg* msg){
+bool filter(prov_entry_t* msg){
   return false;
 }
 
@@ -154,7 +149,6 @@ void log_error(char* err_msg){
 struct provenance_ops ops = {
   .init=&init,
   .filter=&filter,
-  .long_filter=&long_filter,
   .log_unknown_relation=&log_unknown_relation,
   .log_derived=&log_derived,
   .log_generated=&log_generated,