+++ src/provenancelib.c
@@ -23,7 +23,7 @@
 
 #include "provenancelib.h"
 #include "provenanceutils.h"
-#include "libut.h"
+#include "uthash.h"
 
 static inline int __set_boolean(bool value, const char* name){
   int rc;
@@ -93,7 +93,7 @@ declare_get_boolean_fcn(provenance_get_all, PROV_ALL_FILE);
 }
 
 #define declare_self_get_flag(fcn_name, element) bool fcn_name( void ){\
-  union prov_msg self;\
+  union prov_elt self;\
   provenance_self(&self.task_info);\
   return prov_check_flag(&self, element);\
 }
@@ -110,9 +110,8 @@ declare_self_get_flag(provenance_get_propagate, PROPAGATE_BIT);
 int provenance_set_propagate(bool v){
   int err;
   err = __provenance_set_propagate(v);
-  if(err < 0){
+  if(err < 0)
     return err;
-  }
   return provenance_set_tracked(v);
 }
 
@@ -121,14 +120,11 @@ int provenance_set_machine_id(uint32_t v){
   int fd = open(PROV_MACHINE_ID_FILE, O_WRONLY);
 
   if(fd<0)
-  {
     return fd;
-  }
   rc = write(fd, &v, sizeof(uint32_t));
   close(fd);
-  if(rc<0){
+  if(rc<0)
     return rc;
-  }
   return 0;
 }
 
@@ -137,14 +133,37 @@ int provenance_get_machine_id(uint32_t* v){
   int fd = open(PROV_MACHINE_ID_FILE, O_RDONLY);
 
   if(fd<0)
-  {
     return fd;
-  }
   rc = read(fd, v, sizeof(uint32_t));
   close(fd);
-  if(rc<0){
+  if(rc<0)
     return rc;
-  }
+  return 0;
+}
+
+int provenance_set_boot_id(uint32_t v){
+  int rc;
+  int fd = open(PROV_BOOT_ID_FILE, O_WRONLY);
+
+  if(fd<0)
+    return fd;
+  rc = write(fd, &v, sizeof(uint32_t));
+  close(fd);
+  if(rc<0)
+    return rc;
+  return 0;
+}
+
+int provenance_get_boot_id(uint32_t* v){
+  int rc;
+  int fd = open(PROV_BOOT_ID_FILE, O_RDONLY);
+
+  if(fd<0)
+    return fd;
+  rc = read(fd, v, sizeof(uint32_t));
+  close(fd);
+  if(rc<0)
+    return rc;
   return 0;
 }
 
@@ -153,9 +172,7 @@ int provenance_disclose_node(struct disc_node_struct* node){
   int fd = open(PROV_NODE_FILE, O_WRONLY);
 
   if(fd<0)
-  {
     return fd;
-  }
   rc = write(fd, node, sizeof(struct disc_node_struct));
   close(fd);
   return rc;
@@ -166,9 +183,7 @@ int provenance_disclose_relation(struct relation_struct* relation){
   int fd = open(PROV_RELATION_FILE, O_WRONLY);
 
   if(fd<0)
-  {
     return fd;
-  }
   rc = write(fd, relation, sizeof(struct relation_struct));
   close(fd);
   return rc;
@@ -179,18 +194,15 @@ int provenance_self(struct task_prov_struct* self){
   int fd = open(PROV_SELF_FILE, O_RDONLY);
 
   if(fd<0)
-  {
     return fd;
-  }
   rc = read(fd, self, sizeof(struct task_prov_struct));
   close(fd);
   return rc;
 }
 
 bool provenance_is_present(void){
-  if(access(PROV_ENABLE_FILE, F_OK)){ // return 0 if file exists.
+  if(access(PROV_ENABLE_FILE, F_OK)) // return 0 if file exists.
     return false;
-  }
   return true;
 }
 
@@ -200,32 +212,30 @@ int provenance_flush(void){
   int fd = open(PROV_FLUSH_FILE, O_WRONLY);
 
   if(fd<0)
-  {
     return fd;
-  }
   rc = write(fd, &tmp, sizeof(char));
   close(fd);
   return rc;
 }
 
-int provenance_read_file(const char path[PATH_MAX], union prov_msg* inode_info){
-  return getxattr(path, XATTR_NAME_PROVENANCE, inode_info, sizeof(union prov_msg));
+int provenance_read_file(const char path[PATH_MAX], union prov_elt* inode_info){
+  return getxattr(path, XATTR_NAME_PROVENANCE, inode_info, sizeof(union prov_elt));
 }
 
-int fprovenance_read_file(int fd, union prov_msg* inode_info){
-  return fgetxattr(fd, XATTR_NAME_PROVENANCE, inode_info, sizeof(union prov_msg));
+int fprovenance_read_file(int fd, union prov_elt* inode_info){
+  return fgetxattr(fd, XATTR_NAME_PROVENANCE, inode_info, sizeof(union prov_elt));
 }
 
-static inline int __provenance_write_file(const char path[PATH_MAX], union prov_msg* inode_info){
-  return setxattr(path, XATTR_NAME_PROVENANCE, inode_info, sizeof(union prov_msg), 0);
+static inline int __provenance_write_file(const char path[PATH_MAX], union prov_elt* inode_info){
+  return setxattr(path, XATTR_NAME_PROVENANCE, inode_info, sizeof(union prov_elt), 0);
 }
 
-static inline int __fprovenance_write_file(int fd, union prov_msg* inode_info){
-  return fsetxattr(fd, XATTR_NAME_PROVENANCE, inode_info, sizeof(union prov_msg), 0);
+static inline int __fprovenance_write_file(int fd, union prov_elt* inode_info){
+  return fsetxattr(fd, XATTR_NAME_PROVENANCE, inode_info, sizeof(union prov_elt), 0);
 }
 
 static inline int __provenance_set_flags_file(const char path[PATH_MAX], uint8_t bit, bool v){
-  union prov_msg prov;
+  union prov_elt prov;
   int rc;
   rc = provenance_read_file(path, &prov);
   if(rc<0)
@@ -238,7 +248,7 @@ static inline int __provenance_set_flags_file(const char path[PATH_MAX], uint8_t
 }
 
 static inline int __fprovenance_set_flags_file(int fd, uint8_t bit, bool v){
-  union prov_msg prov;
+  union prov_elt prov;
   int rc;
   rc = fprovenance_read_file(fd, &prov);
   if(rc<0)
@@ -269,23 +279,22 @@ declare_fset_file_fcn(__fprovenance_propagate_file, PROPAGATE_BIT);
 int provenance_propagate_file(const char name[PATH_MAX], bool propagate){
   int err;
   err = __provenance_propagate_file(name, propagate);
-  if(err < 0){
+  if(err < 0)
     return err;
-  }
   return provenance_track_file(name, propagate);
 }
 
 int fprovenance_propagate_file(int fd, bool propagate){
   int err;
   err = __fprovenance_propagate_file(fd, propagate);
-  if(err < 0){
+  if(err < 0)
     return err;
-  }
   return fprovenance_track_file(fd, propagate);
 }
 
-int provenance_taint_file(const char path[PATH_MAX], uint64_t taint){
-  union prov_msg prov;
+int provenance_label_file(const char path[PATH_MAX], const char *label){
+  union prov_elt prov;
+  uint64_t taint = generate_label(label);
   int rc;
   rc = provenance_read_file(path, &prov);
   if(rc<0)
@@ -294,8 +303,9 @@ int provenance_taint_file(const char path[PATH_MAX], uint64_t taint){
   return __provenance_write_file(path, &prov);
 }
 
-int fprovenance_taint_file(int fd, uint64_t taint){
-  union prov_msg prov;
+int fprovenance_label_file(int fd, const char *label){
+  union prov_elt prov;
+  uint64_t taint = generate_label(label);
   int rc;
   rc = fprovenance_read_file(fd, &prov);
   if(rc<0)
@@ -304,13 +314,13 @@ int fprovenance_taint_file(int fd, uint64_t taint){
   return __fprovenance_write_file(fd, &prov);
 }
 
-int provenance_taint(uint64_t taint){
+int provenance_label(const char *label){
   struct prov_process_config cfg;
+  uint64_t taint = generate_label(label);
   int rc;
   int fd = open(PROV_SELF_FILE, O_WRONLY);
-  if( fd < 0 ){
+  if( fd < 0 )
     return fd;
-  }
   memset(&cfg, 0, sizeof(struct prov_process_config));
   cfg.op=PROV_SET_TAINT;
   prov_bloom_add(prov_taint(&(cfg.prov)), taint);
@@ -320,19 +330,18 @@ int provenance_taint(uint64_t taint){
   return rc;
 }
 
-int provenance_read_process(uint32_t pid, union prov_msg* process_info){
+int provenance_read_process(uint32_t pid, union prov_elt* process_info){
   struct prov_process_config cfg;
   int rc;
   int fd = open(PROV_PROCESS_FILE, O_RDONLY);
 
-  if( fd < 0 ){
+  if( fd < 0 )
     return fd;
-  }
   cfg.vpid = pid;
 
   rc = read(fd, &cfg, sizeof(struct prov_process_config));
   close(fd);
-  memcpy(process_info, &(cfg.prov), sizeof(union prov_msg));
+  memcpy(process_info, &(cfg.prov), sizeof(union prov_elt));
   return rc;
 }
 
@@ -362,19 +371,18 @@ declare_set_process_fcn(__provenance_propagate_process, PROPAGATE_BIT, PROV_SET_
 int provenance_propagate_process(uint32_t pid, bool propagate){
   int err;
   err = __provenance_propagate_process(pid, propagate);
-  if(err < 0){
+  if(err < 0)
     return err;
-  }
   return provenance_track_process(pid, propagate);
 }
 
-int provenance_taint_process(uint32_t pid, uint64_t taint){
+int provenance_label_process(uint32_t pid, const char *label){
   struct prov_process_config cfg;
+  uint64_t taint = generate_label(label);
   int rc;
   int fd = open(PROV_PROCESS_FILE, O_WRONLY);
-  if( fd < 0 ){
+  if( fd < 0 )
     return fd;
-  }
   memset(&cfg, 0, sizeof(struct prov_process_config));
   cfg.vpid=pid;
   cfg.op=PROV_SET_TAINT;
@@ -393,7 +401,10 @@ union ipaddr{
 static int __param_to_ipv4_filter(const char* param, struct prov_ipv4_filter* filter){
   int err;
   union ipaddr ip;
-  uint32_t a,b,c,d;
+  uint32_t a;
+  uint32_t b;
+  uint32_t c;
+  uint32_t d;
   uint32_t mask;
   uint32_t port;
 
@@ -496,7 +507,9 @@ bool find_entry(uint32_t secid, char* secctx) {
 
 int provenance_secid_to_secctx( uint32_t secid, char* secctx, uint32_t len){
   struct secinfo info;
-  int rc, fd;
+  int rc;
+  int fd;
+
   if( find_entry(secid, secctx) )
     return 0;
   fd = open(PROV_SECCTX, O_RDONLY);
@@ -510,9 +523,8 @@ int provenance_secid_to_secctx( uint32_t secid, char* secctx, uint32_t len){
     secctx[0]='\0';
     return rc;
   }
-  if(len<strlen(info.secctx)){
+  if(len<strlen(info.secctx))
     return -1;
-  }
   strncpy(secctx, info.secctx, len);
   add_entry(secid, secctx);
   return rc;